feat: реализация эндпоинтов category, добавление валидации

+++ main-service/src/main/java/ru/practicum/categories/CategoryController.java
@@ -4,10 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.categories.dto.CategoryDto;
 import ru.practicum.categories.service.CategoryService;
+import ru.practicum.util.exception.RequiredArgsMissingException;
+import ru.practicum.util.exception.ValidationException;
 
 import java.util.List;
 
 import static ru.practicum.categories.dto.CategoryValidator.validateNewCategory;
+import static ru.practicum.util.VariableValidator.validatePageableParams;
 
 @RestController
 @RequiredArgsConstructor
@@ -21,14 +24,22 @@ public class CategoryController {
     }
 
     @DeleteMapping("/admin/categories/{catId}")
-    public void deleteCategory(@PathVariable long catId) {
+    public void deleteCategory(@PathVariable Long catId) {
+        if (catId == null)
+            throw new RequiredArgsMissingException("Required variable is missing", "Path variable 'catId' is missing");
+        if (catId <= 0)
+            throw new ValidationException("Invalid path variable", "Path variable 'catId' should be positive");
         service.deleteCategory(catId);
     }
 
     @PatchMapping("/admin/categories/{catId}")
     public CategoryDto updateCategory(
-            @PathVariable long catId,
+            @PathVariable Long catId,
             @RequestBody CategoryDto categoryDto) {
+        if (catId == null)
+            throw new RequiredArgsMissingException("Required variable is missing", "Path variable 'catId' is missing");
+        if (catId <= 0)
+            throw new ValidationException("Invalid path variable", "Path variable 'catId' should be positive");
         validateNewCategory(categoryDto);
         return service.updateCategory(catId, categoryDto);
     }
@@ -37,11 +48,20 @@ public class CategoryController {
     public List<CategoryDto> getCategoryList(
             @RequestParam(required = false, defaultValue = "0") int from,
             @RequestParam(required = false, defaultValue = "10") int size) {
-        return null;
+        validatePageableParams(from, size);
+        if (from < 0)
+            throw new ValidationException("Invalid page params", "Param 'from' should be positive or zero");
+        if (size < 1)
+            throw new ValidationException("Invalid page params", "Param 'size' should be positive");
+        return service.getCategoryList(from, size);
     }
 
     @GetMapping("/categories/{catId}")
-    public CategoryDto getCategoryById(@PathVariable long catId) {
-        return null;
+    public CategoryDto getCategoryById(@PathVariable Long catId) {
+        if (catId == null)
+            throw new RequiredArgsMissingException("Required variable is missing", "Path variable 'catId' is missing");
+        if (catId <= 0)
+            throw new ValidationException("Invalid path variable", "Path variable 'catId' should be positive");
+        return service.getCategoryById(catId);
     }
 }
